scocca_concept_discovery/requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
